+++ mkosi/backend.py
@@ -210,7 +210,7 @@ class CommandLineArguments:
     skeleton_trees: List[str]
     remove_files: List[str]
     build_script: Optional[str]
-    build_env: List[str]
+    environment: List[str]
     build_sources: Optional[str]
     build_dir: Optional[str]
     include_dir: Optional[str]
@@ -359,11 +359,15 @@ def nspawn_params_for_blockdev_access(args: CommandLineArguments, loopdev: str) 
         "--bind-ro=/dev/block",
         "--bind-ro=/dev/disk",
     ]
+
     for partno in (args.esp_partno, args.bios_partno, args.root_partno, args.xbootldr_partno):
         if partno is not None:
             p = partition(loopdev, partno)
             if os.path.exists(p):
                 params += [f"--bind-ro={p}", f"--property=DeviceAllow={p}"]
+
+    params.extend(f"--setenv={env}" for env in args.environment)
+
     return params